+++ frontend/pages/label/[id].vue
@@ -2,7 +2,7 @@
   import type { DateDetail, Detail } from "~~/components/global/DetailsSection/types";
 
   definePageMeta({
-    layout: "home",
+    middleware: ["auth"],
   });
 
   const route = useRoute();
